requirements.txt
fastapi
uvicorn
torch
numpy
openai-whisper
indic-transliteration
mediapipe
